Clear keychain items when first-time app launch after installation
If the user uninstalls the app, the token stored in the keychain item
will not delete. So that if user reinstalls the app, they will
automatically get into exams list using existing token.
This is fixed by clearing the existing token when the user launches the
app for the first time after reinstall.